dev-requirements.txt
PyYAML
networkx
matplotlib
graphviz